% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/opennhtsa.R
\name{\%>\%}
\alias{\%>\%}
\alias{chain_query}
\title{Pipe operator for chaining together operations.}
\description{
Imported from magrittr, use this to chain together
operations in a natural way.
}
\examples{
\dontrun{
# instead of
a(b(c("hello")), "bob")

# we can write:

c("hello") \\\%>\\\% b() \\\%>\\\% a("bob")

# this also allows for currying common arguments:

my_query <- facility("recalls") \%>\%
              model_year("2010") \%>\%
              vehicle_make("ford") \%>\%
              vehicle_model("f-150") \%>\%
              nhsta_fetch()
}
}

