% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/opennhtsa.R
\name{nhtsa_fetch}
\alias{nhtsa_fetch}
\title{Fetch the given URL as JSON.}
\usage{
nhtsa_fetch(url, debug = FALSE)
}
\value{
data.frame
}
\description{
This uses jsonlite to fetch the URL.  The result is coerced into
a data frame.
}

