% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/opennhtsa.R
\docType{package}
\name{openNHTSA}
\alias{openNHTSA}
\alias{openNHTSA-package}
\title{openNHTSA: A package for interfacing with the U.S. Department of Transportation's National Highway Traffic Safety Administration API}
\description{
This package provides a simple wrapper around the NHTSA API (http://www.nhtsa.gov/webapi/Default.aspx?Recalls/API/83).
Credit for the design and architecture of the wrapper should go to the authors of the openfda wrapper housed here: https://github.com/rOpenHealth/openfda/
Many parts of this wrapper are direct copies of their work and hence this package is released under license GPL v2.
}
\details{
It uses the \code{magrittr} piping interface to simplify
building complex queries.
}
\examples{
# Queries generally have the following format
\dontrun{
facility("recalls") \%>\%
  model_year("2010") \%>\%
  vehicle_make("ford") \%>\%
  vehicle_model("fusion") \%>\%
  nhtsa_fetch()

facility("complaints") \%>\%
  model_year("2010") \%>\%
  vehicle_make("ford") \%>\%
  vehicle_model("fusion") \%>\%
  nhtsa_fetch()
}
}

